move PastClosure to ffront.stages
