+++ src/gt4py/next/otf/stages.py
@@ -17,8 +17,6 @@ from __future__ import annotations
 import dataclasses
 from typing import Any, Generic, Optional, Protocol, TypeVar
 
-from gt4py.next import common
-from gt4py.next.ffront import program_ast as past
 from gt4py.next.iterator import ir as itir
 from gt4py.next.otf import languages
 from gt4py.next.otf.binding import interface
@@ -30,15 +28,6 @@ SettingT = TypeVar("SettingT", bound=languages.LanguageSettings)
 SrcL_co = TypeVar("SrcL_co", bound=languages.LanguageTag, covariant=True)
 TgtL_co = TypeVar("TgtL_co", bound=languages.LanguageTag, covariant=True)
 SettingT_co = TypeVar("SettingT_co", bound=languages.LanguageSettings, covariant=True)
-
-
-@dataclasses.dataclass(frozen=True)
-class PastClosure:
-    closure_vars: dict[str, Any]
-    past_node: past.Program
-    grid_type: Optional[common.GridType]
-    args: tuple[Any, ...]
-    kwargs: dict[str, Any]
 
 
 @dataclasses.dataclass(frozen=True)
